LayoutTests/imported/w3c/web-platform-tests/css/css-scoping/keyframes-001-expected.txt


FAIL @keyframes applies in the shadow tree assert_equals: expected 1 but got 0
